feat(gl): add __on_errored_message__
