+++ executor/testdata/cases/py-core/pitfalls/error_msg.jsonnet
@@ -1,0 +1,8 @@
+local simple = import 'templates/simple.jsonnet';
+simple.run('${jsonnetDir}/error_msg.py') {
+    "calldata": |||
+        {
+            "method": "#error"
+        }
+    |||
+}
