+++ executor/testdata/cases/py-core/pitfalls/error_msg.py
@@ -1,0 +1,7 @@
+# { "Depends": "py-genlayer:test" }
+from genlayer import *
+
+
+class Contract(gl.Contract):
+	def __init__(self):
+		pass
