+++ executor/testdata/cases/py-core/pitfalls/error_msg.stdout
@@ -1,0 +1,1 @@
+executed with `Return("\x00")`
