+++ executor/testdata/cases/py-core/pitfalls/error_msg_overridden.jsonnet
@@ -1,0 +1,11 @@
+local simple = import 'templates/simple.jsonnet';
+simple.run('${jsonnetDir}/error_msg_overridden.py') {
+    "calldata": |||
+        {
+            "method": "#error"
+        }
+    |||,
+    message+: {
+        value: 100
+    }
+}
