+++ executor/testdata/cases/py-core/pitfalls/error_msg_overridden.py
@@ -1,0 +1,11 @@
+# { "Depends": "py-genlayer:test" }
+from genlayer import *
+
+
+class Contract(gl.Contract):
+	def __init__(self):
+		pass
+
+	@gl.public.write.payable
+	def __on_errored_message__(self):
+		print('errored but ok')
