+++ executor/testdata/cases/py-core/pitfalls/error_msg_overridden.stdout
@@ -1,0 +1,2 @@
+errored but ok
+executed with `Return("\x00")`
